requirements.txt
flask
requests
beautifulsoup4
pandas
pytz
python-dotenv
aiohttp